refactor: move room logic to usePeerRoom hook
